Add day 13 part 2.
Remove lots of error-prone index handling in favor of just comparing forwards and backwards for each row, checking for errors.

Took forever to iron out the kinks and correctly interpret the actual problem text, namely that there can only ever be one line to fix, and the correct reflection line in part 2 cannot be the same as part 1. This simplified the point calculation tremendously as there is only ever one correct solution.