Add day 23/A Long Walk.
Was able to solve part 1 recursively pretty quickly,
only had to push up the system recursion limit a bit.
That was unsuitable for part 2, as it was running for
30+ minutes and still finding longer paths. Had to
make a proper graph data structure, as the actual
moving about was consuming almost all cpu time. Some
tree search logic later yielded the correct answer
in about 15 seconds.